// src/app/models/vopportunity.model.ts
import { Timestamp } from "rxjs";

export class VOpportunity {
    id: string;
    title: string;
    description: string;
    photoURL: string;
    starcolor: string;
    isflexible: string;
    fromdate: string;
    todate: string;
    location: string;
    city: string;
    country: string;
    isvirtual: string;
    causeid: string;
    organizationid: string;
    requiredskill: string;
    createdDate: Date;
    uid: string;
}